Fix logging level for one event/message type.
This commit fixes the logging level (should be info) for an event/message type
which didn't have it set, and also adds some more potentially useful details to
the log/event message (e.g., the template ID, etc.), plus removes logging from
where we pull for an instance state, as this information would end up in the
log/event messages (this needs to be revised going forward). Since it only would
be printed as-is without Zerlog's JSON format, it should be easy to filter it
out from the messages considered an audit/event log.

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>